Setting to split concatenated words
Since using this keyboard a major issue i've found is how easy it is
to accidentally write concatenated words where I accidentally press a
letter instead of the space bar. I'm constantly writing c, v, b, n,
and m (bottom row keys) instead of space leading to words like
"hellobthere" or "whatbarenyouvdoingvtonight?"

My proposal is adding a setting which would correct for this specific
issue by trying to split unknown words by the trouble keys and testing
for dictionary values in the split words. In my own testing this works
great and my number of mispells and retypes has gone down drastically.

An alternative fix for this is to enlarge the whole keyboard thus
increasing padding around the space bar, but it affects ergonomics and
what you can read on screen. Another possibility is to enlarge the
space bar but that also affects screen size.

This was not an issue with major competing keyboards as I suspect they
either correct using their own proprietary models, or they do
something similar to this change. I have no intention of using those
keyboards so I believe Heliboard should be able to solve the problem.
I believe this solution is a better enhancement and provides for a
good user experience when enabled.